LayoutTests/media/accessibility-closed-captions-has-aria-owns-expected.txt

Bug 127065: AX: Shadow DOM video player controls menus need aria-owns on the trigger buttons

Does the `aria-owns` on the CC button exist?

PASS captionsButtonARIAOwnsValue is not null



Did the Audio Track menu show up after the CC button is clicked?

PASS captionsTrackMenuHTMLElement is not null



Does the `id` of the menu exist?

PASS closedCaptionsTrackMenuIdValue is not null



Is the `aria-own` on the CC button equal to the `id` on the menu?

PASS captionsButtonARIAOwnsValue is closedCaptionsTrackMenuIdValue
PASS successfullyParsed is true

TEST COMPLETE
